moved ServiceType to models
