+++ services/director-v2/src/simcore_service_director_v2/models/schemas/dynamic_services/service.py
@@ -9,7 +9,20 @@ from models_library.projects_nodes_io import NodeID
 from models_library.services import DYNAMIC_SERVICE_KEY_RE, VERSION_RE
 from pydantic import BaseModel, Field
 
+from ....meta import api_vtag
 from ..constants import UserID
+
+
+@unique
+class ServiceType(Enum):
+    """
+    Used to filter out services spawned by this service in the stack.
+    The version was added to distinguish from the ones spawned by director-v0
+    These values are attached to the dynamic-sidecar and its relative proxy.
+    """
+
+    MAIN = f"main-{api_vtag}"
+    DEPENDENCY = f"dependency-{api_vtag}"
 
 
 class ServiceDetails(BaseModel):
